refactor #149: move loader to atoms folder
